TS: EntriesOnStructure
