--- src/components/Structure/utils.js
@@ -1,8 +1,0 @@
-export const experimentTypes = new Map([
-  ['x-ray', 'X-ray'],
-  ['nmr', 'NMR'],
-  ['all', 'All'],
-  ['em', 'Cryo-EM'],
-]);
-export const formatExperimentType = (type) =>
-  experimentTypes.get((type || '').toLowerCase(), type.toUpperCase());
